rename ActiveTeachersAllTimeQuery -> ActiveTeachersAllTimeCountQuery, reduce return value to a scalar value
